server/requirements.txt
# oaAnsible Server Dependencies
# FastAPI and ASGI server
fastapi>=0.104.0
uvicorn[standard]>=0.24.0

# Authentication and security
pyjwt>=2.8.0
passlib[bcrypt]>=1.7.4
python-multipart>=0.0.6

# HTTP client for dashboard integration
httpx>=0.25.0

# Database (SQLite with async support)
aiosqlite>=0.19.0

# YAML parsing for Ansible inventories
pyyaml>=6.0.1

# Background tasks and async support
celery>=5.3.0  # Optional: for more robust job queuing
redis>=5.0.0   # Optional: for job queue and caching

# Monitoring and logging
prometheus-client>=0.19.0
structlog>=23.2.0

# Development and testing
pytest>=7.4.0
pytest-asyncio>=0.21.0
pytest-cov>=4.1.0
httpx>=0.25.0  # For testing

# Type hints and validation
pydantic>=2.5.0
typing-extensions>=4.8.0

# CORS support
fastapi-cors>=0.0.6

# Rate limiting
slowapi>=0.1.9

# WebSocket support for real-time updates
websockets>=12.0

# Configuration management
python-dotenv>=1.0.0

# Process management
psutil>=5.9.0